+++ packages/convex-client/src/index.ts
@@ -1,5 +1,7 @@
 export { Convex } from './convex.js';
 export { KeyPair } from './KeyPair.js';
+export { type Signer } from './Signer.js';
+export { KeyPairSigner } from './KeyPairSigner.js';
 export * from './types.js';
 export * from './crypto.js';
 export * from './identicon.js';